requirements.txt
gradio
PyPDF2
sentence-transformers
scikit-learn
langchain-google-genai
python-dotenv